% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/andrast.R
\name{andrast}
\alias{andrast}
\title{INteresction (AND) of catagories in a multi-layer raster/rast object}
\usage{
andrast(x, nmax_cat = 100, name = "andrast", class_number = FALSE, ...)
}
\arguments{
\item{x}{a SpatRast object}

\item{nmax_cat}{maximum number of catagories per layer}

\item{name}{output raster name}

\item{class_number}{logical}

\item{...}{further arguments}
}
\description{
INteresction (AND) of catagories in a multi-layer raster/rast object
}
\examples{

library(terra)
library(leaflet)
elev <- rast(system.file("ex/elev_vinschgau.tif",package="terra"))
names(elev) <- "elev"
elev_rcl <- c(0,1000,2000,3000,4000) ##c(0,500,1500,2000,2500,3000,3500)
elevb <- classify(elev,rcl=elev_rcl)
#######
#######
naspect <- 4 ##8
aspect <- terrain(elev,"aspect")
aspect[aspect>360*(1-1/(naspect*2))] <- aspect[aspect>360*(1-1/(naspect*2))]-360
aspect_rcl <- (0:naspect)/naspect*360-360/(naspect*2)
aspectb <- classify(aspect,rcl=aspect_rcl)
#######
#######
laplacian <- laplacian(elev)
laplacian_rcl <- c(-Inf,-0.01,0.01,Inf)
laplacianb <- classify(laplacian,rcl=laplacian_rcl)

x <- c(elevb,aspectb,laplacianb)

out <- andrast(x)
out <- andrast(x,class_number=TRUE)
plet(out,col=rainbow(80),tiles="OpenTopoMap")

 ## TO DO 
}
