% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_grid.R
\name{make_grid}
\alias{make_grid}
\title{Make Grid / Create Grid}
\usage{
make_grid(x, fact = 10, ...)
}
\arguments{
\item{x}{elevation / terrain digital model (DTM) , a SpatRast object}

\item{fact, ...}{furhrer agruments for \code{\link[terra]{aggregate}}}
}
\description{
Make Grid / Create Grid
}
\examples{

elev <- rast(system.file("ex/elev_vinschgau.tif",package="terra"))

out <- make_grid(elev)

plot(elev)
plot(out,add=TRUE)

}
\seealso{
\code{\link[terra]{aggregate}}
}
